Add new Invoice Status, `INVALID`. This status represents an invoice that has the wrong currency. Therefore, it cannot be processed until the currency is fixed.
Enable to update the invoice status (at service & dal level).

Create REST endpoints to update an invoice and to run the billing operation. This enables the users to not rely only in the scheduling (e.g.: enable running in the middle of the month, or after fixing erroneous bills).

Set status as PAID for successful charges in the BillingService. Added a more robust exception treatment to include Currency Mismatch and Customer Not Found.

Fix location of method `fetchInvoicesGroupedByClient` - from `CustomerService` to `InvoicesService`.